// xekuii-web/src/generated/routes.generated.tsx
// Auto-generated by XekuII.Generator - DO NOT EDIT
// Generated at: 2026-02-06 22:03:33

import type { RouteComponent } from "./route-types";
import { ProductListPage } from "./pages/product/ProductListPage.generated";
import { ProductFormPage } from "./pages/product/ProductFormPage.generated";
import { ProductDetailPage } from "./pages/product/ProductDetailPage.generated";
import { PurchaseOrderListPage } from "./pages/purchase-order/PurchaseOrderListPage.generated";
import { PurchaseOrderFormPage } from "./pages/purchase-order/PurchaseOrderFormPage.generated";
import { PurchaseOrderDetailPage } from "./pages/purchase-order/PurchaseOrderDetailPage.generated";
import { PurchaseOrderItemListPage } from "./pages/purchase-order-item/PurchaseOrderItemListPage.generated";
import { PurchaseOrderItemFormPage } from "./pages/purchase-order-item/PurchaseOrderItemFormPage.generated";
import { PurchaseOrderItemDetailPage } from "./pages/purchase-order-item/PurchaseOrderItemDetailPage.generated";
import { SupplierListPage } from "./pages/supplier/SupplierListPage.generated";
import { SupplierFormPage } from "./pages/supplier/SupplierFormPage.generated";
import { SupplierDetailPage } from "./pages/supplier/SupplierDetailPage.generated";

export const generatedRoutes: RouteComponent[] = [
  { path: "/products", component: ProductListPage },
  { path: "/products/new", component: ProductFormPage },
  { path: "/products/$id", component: ProductDetailPage },
  { path: "/products/$id/edit", component: ProductFormPage },
  { path: "/purchase-orders", component: PurchaseOrderListPage },
  { path: "/purchase-orders/new", component: PurchaseOrderFormPage },
  { path: "/purchase-orders/$id", component: PurchaseOrderDetailPage },
  { path: "/purchase-orders/$id/edit", component: PurchaseOrderFormPage },
  { path: "/purchase-order-items", component: PurchaseOrderItemListPage },
  { path: "/purchase-order-items/new", component: PurchaseOrderItemFormPage },
  { path: "/purchase-order-items/$id", component: PurchaseOrderItemDetailPage },
  { path: "/purchase-order-items/$id/edit", component: PurchaseOrderItemFormPage },
  { path: "/suppliers", component: SupplierListPage },
  { path: "/suppliers/new", component: SupplierFormPage },
  { path: "/suppliers/$id", component: SupplierDetailPage },
  { path: "/suppliers/$id/edit", component: SupplierFormPage },
];